+++ lib/src/dataflow/tuple.dart
@@ -51,22 +51,22 @@ class Aes {
   List<Offset> position;
 
   /// The shape of the tuple.
-  final Shape shape;
+  Shape shape;
 
   /// The color of the tuple.
-  final Color? color;
+  Color? color;
 
   /// The gradient of the tuple.
-  final Gradient? gradient;
+  Gradient? gradient;
 
   /// The shadow elevation of the tuple.
-  final double? elevation;
+  double? elevation;
 
   /// The label of the tuple.
-  final Label? label;
+  Label? label;
 
   /// The size of the tuple.
-  final double? size;
+  double? size;
 
   /// The represent point of [position] points.
   Offset get representPoint => shape.representPoint(position);