+++ lib/src/shape/custom.dart
@@ -3,22 +3,10 @@ import 'dart:ui';
 import 'package:graphic/src/coord/coord.dart';
 import 'package:graphic/src/coord/rect.dart';
 import 'package:graphic/src/dataflow/tuple.dart';
-import 'package:graphic/src/geom/custom.dart';
 import 'package:graphic/src/graffiti/figure.dart';
 
 import 'util/render_basic_item.dart';
 import 'shape.dart';
-
-/// The shape for the custom element.
-///
-/// See also:
-///
-/// - [CustomElement], which this shape is for.
-abstract class CustomShape extends Shape {
-  @override
-  double get defaultSize => throw UnimplementedError(
-      'Designate default size in your own custom shape');
-}
 
 /// A candle stick shape.
 ///
@@ -33,7 +21,7 @@ abstract class CustomShape extends Shape {
 /// ** We insist that the price of a subject matter of investment is determined
 /// by its intrinsic value. Too much attention to the short-term fluctuations in
 /// prices is harmful. Thus a candlestick chart may misslead your investment decision.**
-class CandlestickShape extends CustomShape {
+class CandlestickShape extends Shape {
   /// Creates a candle stick shape.
   CandlestickShape({
     this.hollow = true,