+++ lib/src/shape/shape.dart
@@ -1,6 +1,8 @@
 import 'dart:ui';
 
+import 'package:graphic/src/common/customizable_spec.dart';
 import 'package:graphic/src/dataflow/tuple.dart';
+import 'package:graphic/src/geom/custom.dart';
 import 'package:graphic/src/graffiti/figure.dart';
 import 'package:flutter/foundation.dart';
 import 'package:graphic/src/coord/coord.dart';
@@ -11,8 +13,10 @@ import 'package:graphic/src/coord/coord.dart';
 /// the key of painting geometry elements. Besides, the shape it self is an aesthetic
 /// attribute in Grammar of Graphics.
 ///
-/// Customizing subclasses of shape extenses chart types.
-abstract class Shape {
+/// Shapes could be customized by extending its subclasses of different geometory
+/// types, or directly extend this class for the [CustomElement]. Customizing shapes
+/// extenses chart types.
+abstract class Shape extends CustomizableSpec {
   /// Renders the whole group of tuples.
   ///
   /// The tuples are rendered in groups. the [Aes.shape] of the first tuple of a
@@ -44,16 +48,4 @@ abstract class Shape {
   ///
   /// Usually the represent point is the last one.
   Offset representPoint(List<Offset> position) => position.last;
-
-  /// Checks the equlity of two shapes.
-  ///
-  /// Because the shape is a functional class in design, they should be equal if
-  /// they are of the same type and has same properties.
-  ///
-  /// It is used by [==]. This method forces a equlity definition in customizing.
-  @protected
-  bool equalTo(Object other);
-
-  @override
-  bool operator ==(Object other) => this.equalTo(other);
 }